+++ cli/cmd_diff.go
@@ -18,8 +18,7 @@ import (
 
 var OptDiffNumLines = options.NewInt(
 	"diff.lines",
-	"unified",
-	'U',
+	&options.Flag{Name: "unified", Short: 'U'},
 	3,
 	"Generate diffs with <n> lines of context",
 	`Generate diffs with <n> lines of context, where n >= 0.`,
@@ -28,8 +27,7 @@ var OptDiffNumLines = options.NewInt(
 
 var OptDiffDataFormat = format.NewOpt(
 	"diff.data.format",
-	"format",
-	'f',
+	&options.Flag{Name: "format", Short: 'f'},
 	format.Text,
 	func(f format.Format) error {
 		switch f { //nolint:exhaustive
@@ -187,7 +185,7 @@ The default (3) can be changed via:
 	}
 
 	panicOn(cmd.RegisterFlagCompletionFunc(
-		OptDiffDataFormat.Flag(),
+		OptDiffDataFormat.Flag().Name,
 		completeStrings(-1, stringz.Strings(diffFormats)...),
 	))