+++ cli/output/format/opt.go
@@ -9,10 +9,10 @@ var _ options.Opt = Opt{}
 
 // NewOpt returns a new format.Opt instance. If validFn is non-nil, it
 // is executed against possible values.
-func NewOpt(key, flag string, short rune, defaultVal Format,
+func NewOpt(key string, flag *options.Flag, defaultVal Format,
 	validFn func(Format) error, usage, help string,
 ) Opt {
-	opt := options.NewBaseOpt(key, flag, short, usage, help, options.TagOutput)
+	opt := options.NewBaseOpt(key, flag, usage, help, options.TagOutput)
 	return Opt{BaseOpt: opt, defaultVal: defaultVal, validFn: validFn}
 }